// call-center-return/packages/return-shared/src/lib/state/adjust-pricing/adjust-pricing.interface.ts
/*
 * IBM Confidential
 * OCO Source Materials
 * 5737-D18, 5725-D10
 *
 * (C) Copyright International Business Machines Corp. 2022, 2023
 *
 * The source code for this program is not published or otherwise divested
 * of its trade secrets, irrespective of what has been deposited with the
 * U.S. Copyright Office.
 */

export interface AdjustPricingState {
    chargeCategoryList: Array<any>;
    chargeExchangeCategoryList: Array<any>;
    chargeNameList: Array<any>;
    chargeExchangeNameList: Array<any>;
    appliedCharges: any;
    chargeErrorMessage: string;
    couponList: Array<any>;
    changeOrder: any;
}

export const INITIAL_ADJUST_PRICING_STATE: AdjustPricingState = { chargeCategoryList: [], chargeExchangeCategoryList: [], chargeNameList: [], chargeExchangeNameList: [],  appliedCharges: null, chargeErrorMessage: null, couponList:[], changeOrder: null };